fix(front): algolia hits
