+++ frontend/components/searchMenu/SuggestList.tsx
@@ -1,30 +1,35 @@
+import { VFC } from 'react'
 import LocationOn from '@components/icons/LocationOn'
 
-type Props = {
-  click: (propertyId: number) => void
+import { Hit, HitsProvided } from 'react-instantsearch-core'
+import { connectHits } from 'react-instantsearch-dom'
+
+type HitDoc = {
+  objectID: string
+  city: string
+  country: string
 }
 
-const SuggestList = ({ click }: Props) => {
-  // ほんとは繰り返し。今はレイアウト用に暫定
+type Props = HitsProvided<Hit<HitDoc>> & {
+  click: (city: string, country: string) => void
+}
+
+const SuggestList: VFC<Props> = ({ hits, click }) => {
   return (
     <ul>
-      <li className="flex cursor-pointer hover:font-bold" onClick={() => click(1)}>
-        <LocationOn width="22px" height="22px" className="fill-gray-4F4F4F" />
-        <p className="ml-2.5 font-Mulish text-sm">Helsinki, Finland</p>
-      </li>
-      <li className="mt-9 flex cursor-pointer hover:font-bold">
-        <LocationOn width="22px" height="22px" className="fill-gray-4F4F4F" />
-        <p className="ml-2.5 font-Mulish text-sm">Turku, Finland</p>
-      </li>
-      <li className="mt-9 flex cursor-pointer hover:font-bold">
-        <LocationOn width="22px" height="22px" className="fill-gray-4F4F4F" />
-        <p className="ml-2.5 font-Mulish text-sm">Oulu, Finland</p>
-      </li>
-      <li className="mt-9 flex cursor-pointer hover:font-bold">
-        <LocationOn width="22px" height="22px" className="fill-gray-4F4F4F" />
-        <p className="ml-2.5 font-Mulish text-sm">Vaasa, Finland</p>
-      </li>
+      {hits!.map((hit: HitDoc) => (
+        <li
+          className="flex cursor-pointer hover:font-bold"
+          onClick={() => click(hit.city, hit.country)}
+          key={hit.objectID}
+        >
+          <LocationOn width="22px" height="22px" className="fill-gray-4F4F4F" />
+          <p className="ml-2.5 font-Mulish text-sm">
+            {hit.city}, {hit.country}
+          </p>
+        </li>
+      ))}
     </ul>
   )
 }
-export default SuggestList
+export default connectHits<Props, Hit<HitDoc>>(SuggestList)
